requirements.txt
# Core ML
torch>=2.1.0
transformers>=4.38.0
datasets>=2.16.0
accelerate>=0.25.0
peft>=0.7.0
trl>=0.7.6
bitsandbytes>=0.41.0

# DeepSpeed
deepspeed>=0.12.0

# Evaluation & Analysis
evaluate>=0.4.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
numpy>=1.24.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
wandb>=0.15.0
huggingface_hub>=0.19.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# API (for preference generation)
requests>=2.31.0
openai>=1.0.0
